More reasonable inertia values. (#58)
To cite the URDF tutorial on physical properties (http://docs.ros.org/en/galactic/Tutorials/URDF/Adding-Physical-and-Collision-Properties-to-a-URDF-Model.html):
"If unsure what to put, a matrix with ixx/iyy/izz=1e-3 or smaller is often a reasonable default for a mid-sized link (it corresponds to a box of 0.1 m side length with a mass of 0.6 kg). The identity matrix is a particularly bad choice, since it is often much too high (it corresponds to a box of 0.1 m side length with a mass of 600 kg!)."
For pedagogic reasons, the urdf_tutorial examples should not do the opposite of this.